# Conflicts:
#	Args/Resources/Lib/Args-Core.lib
#	Args/Resources/Lib/Args-Input.lib
#	Args/Resources/Lib/Args-Rendering.lib
#	Args/Resources/Lib/Args-Window.lib

+++ Args/Args-Engine/Source/Networking/Client.cpp
@@ -42,7 +42,7 @@ void Client::start()
 	gets_s(message, SOCKET_BUFFER_SIZE);
 
 	int flags = 0;
-	if (sendto(sock, message, strlen(message), flags, (SOCKADDR*)&server_address, sizeof(server_address)) == SOCKET_ERROR)
+	if (sendto(sock, message, (int)strlen(message), flags, (SOCKADDR*)&server_address, sizeof(server_address)) == SOCKET_ERROR)
 	{
 		printf("sendto failed: %d", WSAGetLastError());
 		return;
